TACO-176 Separate Apstag from A9Provider, make it work for logged in users
